requirements.txt
absl-py==0.7.1
astor==0.8.0
atari-py==0.1.15
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
colorlog==4.0.2
Cython==0.29.10
decorator==4.4.0
filelock==3.0.12
flatbuffers==1.11
funcsigs==1.0.2
future==0.17.1
gast==0.2.2
glfw==1.8.1
GPUtil==1.4.0
grpcio==1.21.1
gym==0.12.5
h5py==2.9.0
idna==2.8
imageio==2.5.0
importlib-metadata==0.18
ipdb==0.12
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
lockfile==0.12.2
lz4==2.1.10
Markdown==3.1.1
mock==3.0.5
more-itertools==7.0.0
numpy==1.17.3
packaging==19.0
pandas==0.24.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
pluggy==0.12.0
prompt-toolkit==2.0.9
protobuf==3.8.0
psutil==5.6.3
ptyprocess==0.6.0
py==1.8.0
py-spy==0.1.11
pycparser==2.19
pyglet==1.3.2
Pygments==2.4.2
pynput==1.4.2
pyparsing==2.4.0
pytest==4.6.3
python-dateutil==2.8.0
python-xlib==0.25
pytz==2019.1
PyYAML==5.1.1
redis==3.2.1
scipy==1.3.0
tensorboardX
tensorflow-gpu==1.15.0
termcolor==1.1.0
traitlets==4.3.2
typing==3.6.6
urllib3==1.25.3
git+https://github.com/alex-petrenko/ViZDoom@doom_bot_project#egg=vizdoom
wcwidth==0.1.7
Werkzeug==0.15.4
yapf==0.27.0
zipp==0.5.1
matplotlib